Force generated points to be integers
Apparently this behavior is different with different python versions.
Force it to be an integer so the rest of the stack can handle it.

+++ y2020/vision/tools/python_code/target_definition_ui.py
@@ -91,7 +91,7 @@ _on_click_id = None
 
 
 def on_click(event, fig, ax, timer, text, td):
-    point2d = [round(event.xdata), round(event.ydata)]
+    point2d = [int(round(event.xdata)), int(round(event.ydata))]
     # They clicked on the target point after finishing polygon points
     if len(td.polygon_list[0]) == len(td.polygon_list_3d[0]):
         # Disconnect the on click so that nothing gets called if the user
